Moved HeadingFinder to PlatoonService
